refactor(image-generation): move image generation files out of llm
