+++ front/lib/api/actions/servers/image_generation/imageGeneration.ts
