+++ front/lib/api/actions/servers/image_generation/tools/index.ts
@@ -5,6 +5,7 @@ import type {
 } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { buildTools } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import type { ToolContext } from "@app/lib/actions/types";
+import { getImageGenerationLLM } from "@app/lib/api/actions/servers/image_generation/getImageGenerationLLM";
 import {
   checkImageGenerationRateLimit,
   computeImageGenerationCostDetails,
@@ -13,12 +14,11 @@ import {
   trackTokenUsage,
   uploadAndFormatImageResponse,
 } from "@app/lib/api/actions/servers/image_generation/helpers";
-import { IMAGE_GENERATION_TOOLS_METADATA } from "@app/lib/api/actions/servers/image_generation/metadata";
-import { getImageGenerationLLM } from "@app/lib/api/llm/getImageGenerationLLM";
 import type {
   ImageGenerationInput,
   ReferenceImageFile,
-} from "@app/lib/api/llm/imageGeneration";
+} from "@app/lib/api/actions/servers/image_generation/imageGeneration";
+import { IMAGE_GENERATION_TOOLS_METADATA } from "@app/lib/api/actions/servers/image_generation/metadata";
 import type { Authenticator } from "@app/lib/auth";
 import { getStatsDClient } from "@app/lib/utils/statsd";
 import logger from "@app/logger/logger";
